[timekeeper] Update inspect to work in root(covariance).
We were forced to work with square root of covariance in Cobalt but it
is generally better - more intuitive and expressable as a zx::Duration.

Since we already need to perform the square root for cobalt, this CL
moves inspect over to also reporting the square root.

Change-Id: I595dbfdd28f7a0b704400940b7aec857883e3315
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444612
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>