[audio_core][tests] Fix windowing function
The Tukey windowing function is used during the clock-synchronization
integration tests. The function is passed an AudioBufferSlice,
processes it and returns an AudioBuffer with the results. This CL
fixes how the processing uses buffer vector indexing so that the
correct AudioBuffer is returned.

Test: test-only, audio-clocks-pipeline-test

Change-Id: I6ec2c5f050b548de785bf2c2f9bf9bafd67dd649
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445641
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>