[testsharder][testrunner] stop on failure for multiply shards
The main utility of multiply is to check whether a test reliably
passes. Once a test has failed once, we know it does not reliably pass.
We can save bot and developer time by exiting early.

This new behavior is not very important and maybe even counter to
expectations for manual uses of multiply shards, but it will become
very important one we automatically multiply affected tests. The
affected tests are the ones most likely to deterministically fail, and
it's a big waste of time to run a test many times when it fails
deterministically.

If we get requests for the old "Keep Going" behavior, we can design a
way to bring that back for manual uses.

Bug: 63143
Change-Id: I2b14aeedf2861132610cac6328e97efa280a7f94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444926
Testability-Review: Gary Miguel <garymm@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>