[components] Add logging integration test
Verify that when there is a component which uses a LogSink
we see a capability route for it. Since the component never tries
to use the LogSink itself we assume this is because component
manager requested the capability on behalf of the component.

Bug: 63000

Change-Id: Ieca62a125d127948f78f3394b72b79271222420d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445635
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Testability-Review: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>