[zbitl] Introduce zbitl::Image
This change introduces zbitl::Image, which represents a "view" that also
permits operations that modify underlying ZBI (we conveniently look the
other way from View<>::iterator::EditHeader). In particular, we
introduce two versions of an operation to append new items: a partial,
more general you-can-write-the-payload-later version and a one-shot
one in which the provided header and data are written to the underlying
storage up front. (The latter is only efficient/sensible when the
desired payload is already in an simple, in-memory format.

Bug: 57244
Change-Id: Ia4dc2ed9d70e8276dc0e8f84a22ac63f9a449fbf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435260
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>