[wlan][policy] Clean up existing connection on new Connect()
When a new Connect() request comes in to the interface manager, we may
re-use an existing client interface. If this client interface had
an existing state machine, we should either:
- re-use the existing state machine and take no further action
- clean up the existing state machine

Prior to this change, the existing state machine was unconditionally
dropped without normal clean-up, which meant there would never be an
update to listeners to notify them that the connection maintained by
that state machine was being terminated. This was largely invisible
when the new connection was for the same network as the prior connection.

Many revisions ago, the deduping of identical Connect() request was
handled by the client state machine itself. With the refactor that made
client state machines particular to one single network, this deduping
functionality was lost and is now recreated in this CL.

Bug: 63086
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: Ie700606861db08c053faf2505424f93655660303
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444955
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Testability-Review: Kevin Sakuma <sakuma@google.com>