[blobfs] Add oldest_revision to superblock
This implements the policy in //src/storage/docs/versioning.md for
tracking the oldest revision of the software that has touched the data.
The current revision is set to 1 so we can differentiate headers
produced before this change. The format is backwards-compatible with old
versions of the driver.

Similarly to was done for FVM in fxrev.dev/442564, rename the `version`
field to `format_version`.

Tested: New unit tests.
Bug: 62593
Change-Id: I3091fccbd0076db3cf388cddf59d97ccaded2937
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444923
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Testability-Review: Brett Wilson <brettw@google.com>