[weave] Refresh end points when the interface goes online/offline.
Whenever a device connects to an AP, the wifi interface acquires an
IP address and goes to online state. When the device disconnects from
an AP, the wifi interface loses IP and goes to offline state. In both
cases, refresh the weave endpoints so that stale endpoints are closed
and new endpoints are be created. This ensures the endpoints are updated
according to network state.

Bug: 62872
Test: fx test weavestack-adaptation-unittests.

Change-Id: Idca08366c9ebc5b5621081f7dbe7b5588adf896a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443956
Commit-Queue: Sundarajan Srinivasan <sundarsrini@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>