[check-licenses] Adding missing license patterns
This CL modifies some of our patterns, and adds many more
to reduce the number of files check-licenses fails to parse
in the fuchsia tree.

This also adds a list of "remaining_unlicensed_files" to main.go.
It's a large list of mostly manual work that we will burn down
as time permits.

We'd like to get the verification enabled, to ensure new unlicensed
files aren't added to the tree, and that list allows us to flip
the switch sooner.

Change-Id: I039bfe80811efcc23c605f477e93b1262fccd337
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445643
Testability-Review: Francois Rousseau <frousseau@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Sean Cuff <seancuff@google.com>