[timekeeper] Refactor clock correction.
This is a precursor to considering skewing the clock.

Behavior is close but not identical with this change. There are a few
corner cases that lead to a slightly different result, e.g. a correction
of less than 1ms previously would have skipped writing to the RTC but
now it doesn't.

There are also more explicit panics for the userspace clock failure cases
which previously led to a log error and skipping to the next iteration.
This seems reasonable given the deterministic clock failure modes and
the current importance of the userspace clock to timekeeper's role.

Change-Id: Icb25a73c1b0005eafdfee7e355c5ec4e1b6baf93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445307
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>