[bootserver] Fix image paths to be relative to image manifest.
If using the images on the filesystem, bootserver would use the relative
paths of the images but make it relative to the current working
directory. This wasn't a problem because in our use cases, the image
manifest and images were always located in the directory that bootserver
was called from. However, this change will make it so bootserver can be
called from anywhere with the path to the image manifest and the image
paths defined in the manifest only need to be relative to the directory
of the manifest.

Change-Id: Ic532245d1a28a298160bcf5930dbb358411c836a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445994
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>