[omaha-client] Reboot sooner if a user asks for an UpdateCheck
If a user initiates an update check in the middle of an update check,
or while we're waiting for a reboot, treat the original update check
as if it was user-initiated. This "updgrades" all update checks to act
as if they were user-initiated if a user asks for a check while
they're running.

Test: added a StateMachine test for a user-initiated update check
while waiting for reboot, and a test for a user-initiated update check
while applying an update (both cases in which we can upgrade in flight
updates).

MULTIPLY: omaha-client-lib-tests

Fixed: 62821
Change-Id: Id3ffb781953248ec6c580507da10554f5318cb80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444925
Commit-Queue: John Wittrock <wittrock@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Testability-Review: Aaron Wood <aaronwood@google.com>