[kernel][sched] Fix precedence edge in thread migration.
Fix an edge case in the precedence of thread migration and time
slice expiration conditions. When an action occurs that requires
thread migration, such as setting a thread's affinity to exclude
the CPU it is running on, a reschedule is required to carry out
the migration. If the expiration of the time slice of the thread
coincides with the need to migrate, the reschedule will first
return the thread to the run queue of the current CPU, when it
should instead migrate immediately to a different CPU.

The kernel test trylock_or_cancel_canceled (k ut timer) is an
example of a situation that can result in a deadlock because of
the thread returning to the run queue instead of migrating as soon
as possible.

Bug: 62970
Test: Multiple instances of QEMU+KVM running k ut -r 500000 timer.
Change-Id: If0082d1186769a475ec5e8e0da31f9a0b256f83f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445815
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>