[kernel][vm] Guard VmCowpages::options_ with lock
The options_ member is currently always accessed with the lock being
held so this change is semantically a no-op, although some code
juggling is needed to convince the static analysis.

Motivation for this is that later on we may want to have nodes
transition to and from different states (such as hidden and not hidden)
and so by guarding options_ now we ensure that no unlocked usages are
added in the future.

Change-Id: I7913587570f4bfd6849ef54c0d9b5d9e1fb33e0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437349
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>