[sfw] Relay InputDeviceRegistry requests to Session
The `fuchsia.input.injection.InputDeviceRegistry` protocol
is implemented by a Session (which is a V2 component),
but needs to be accessible to V1 components.

Normally, V2 components are made available to V1 components
using capability routing. However, that only works for
static children of a component, and the Session is
a dynamic child of the session manager.

Thus, to make `InputDeviceRegistry` available to V1 components
(such as the `input` command-line tool and `sl4f`), the session
manager relays `InputDeviceRegistry` requests to the Session.

Note that `InputDeviceRegistry` protocol is used only to connect
fake devices to the input pipeline, _not_ to inject input
events into the pipeline. Hence, this relay does not introduce
latency on the data path.

Bug: 62363
Test: fx test session_manager_tests

Change-Id: Ia26b948eb4691a83fcccdcdb8bc404b300356252
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445309
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>