[kernel][lib/arch] Rename test files
The "x86-" prefix does not add any real information, and lends itself to
an eyesore of a naming scheme as more tests are added. Moreover, we
qualify the existing CPUID tests as "basic", to make room for other
test files (soon to be added) that will introduce other CPUID-related
content.

We do this now, so as not to complicate subsequent changes.

Bug: 61093
Change-Id: Ia1e218d0470a0f413f781a45648cefa396a13f16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446414
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>