[wlan][policy] remember active scan connections
This change will remember when a network config was used to connect to a
network after the network required an active scan to be seen. This
reduces reconnect time for hidden networks and increases the probability
of seeing a hidden network in a scan.

Since we can now have conflicting data indicating hidden network or not
hidden network (ie seen in passive scan and connected to after active
scan), we will use whatever has happened most recently.

HiddenStats is removed for now because this calculation of hidden
probability does not actually need that information. It will be added
back if/when it is needed.

Multiply: wlancfg-tests
Test: Modified and added tests:
	fx test wlancfg-tests
Bug: 62646
Change-Id: I88e83bcaa2ad184bd1c5a5c0a7e02e8b51682795
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/445556
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Testability-Review: Marc Khouri <mnck@google.com>