[audio][lib] Add Buffer::Append, fix Slice::begin|end
This CL adds the ability to add an AudioBuffer or AudioBufferSlice
to the end of an AudioBuffer. Implementation was blocked by a bug
in AudioBufferSlice::begin() and end(), which are now also fixed.
Unittests are added for the Append methods, and existing unittests
are expanded (they now fail without the fixes).

Test: AudioBuffer and AudioBufferSlice are test-only

Change-Id: I34181594191be21cd20890d1344aac8e6006a4b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446550
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>