[wlan][sme] Always forward Stop AP request
Right now if a request to start AP times out, we can get into a weird
state where we can neither stop the AP (because we do not forward
stop request in Idle state) nor get into started state (because we
process ALREADY_STARTED response as an error).

To fix this case, we always forward the Stop AP request in Idle state,
though with a blank SSID. The lower layer does not use the SSID in stop
request, so it does not cause an issue.

Bug: 61097
Test:
1. fx test wlan-sme-tests wlan-mlme-tests
2. fx test fuchsia-pkg://fuchsia.com/wlan-hw-sim-test#meta/sim-client-vs-sim-ap.cmx
3. Manually verified that AP can recover after failing to start the
   first time

Change-Id: I8dcf23d81b630fbb839de1ae3189b856c2b24c5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436078
Commit-Queue: Kiet Tran <kiettran@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Testability-Review: Kevin Sakuma <sakuma@google.com>