[debugger] Fix crash on build ID error.
The symbol server code would crash attempting to extract an invalid
substring if the build ID was empty or very short. This change avoids
the crash and reports an error. This does not solve the problem where
build IDs are wrong in some cases.

Fixes some cases in the symbol server code where an error would be
reported but the corresponding Err() message was unset. Clean up some
return values and use the ErrOr template to make it more difficult to
make this kind of mistake.

Bug: 63298
Change-Id: I68c619adce0b027cb1386dd52164ae6696da319f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446595
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>