[system-updater] Simplify boot metadata changes.
Before writing updates, the system-updater needs to put boot metadata
into a happy state. The happy state is "current is active and
alternate is unbootable". Before this change, the system-updater took
a number of steps that depended on whether current was already active,
and whether current was marked successful or not.

This change simplifies that substantially. The new algorithm is:

1. If the current configuration is not marked successful, abort the
   update.
2. Mark the alternate configuration unbootable.

Put together, this ensures that current is both active and successful,
and that alternate is unbootable. The "cost" of this simplicitly is
that the update fails if current isn't marked as successful, but
that's aligned with our stance than unhealthy systems shouldn't OTA.

This change deletes a lot of code and tests, largely because the new
algorithm is completely agnostic to which configuration is "active" at
the start of the update.

Bug: 62958
Change-Id: Idfd96709fbd485132f8b1f04143f6f484c34c3ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441514
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>
Testability-Review: Ben Keller <galbanum@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>