[system-health-check] Make system-health-check more testable.
Currently, system_health_check::set_active_configuration_healthy gets
a handle to the PaverService via the envionrment, which is why we have
a whole complicated integration test that wraps the library in a
binary just to pass the mock paver in.

Instead, set_active_configuration_healthy now accepts a
BootManagerProxy so it's much easier to test, and no longer has
implicit dependencies on the environment. Accordingly, this also
changes system-update-checker and omaha-client to explicitly provision
the paver for the system-health-check library.

All the test that were in mark-active-configuration-healthy-tests are
now unit tests.

Context: while I think this is a positive change regardless, I've
mostly put it into its own CL in order to reduce the line count of a
much bigger CL to fix fxb/51480 (see next in chain).

Bug: 51480
Change-Id: Ie28d1deee4297d681541185eb71161eabf2de51e


Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442995
Commit-Queue: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: Zach Kirschenbaum <zkbaum@google.com>
Testability-Review: Zach Kirschenbaum <zkbaum@google.com>