[bt][rfcomm] Cancel pending channel requests if establishment fails
If the Session terminates, or a specific DLC is not able to be
established, notify clients of the cancellation. This addresses the
issue of a profile client requesting an RFCOMM connection (using
Profile.Connect), the peer either disconnecting or rejecting the
request, and the connection request remaining hanging. In such cases, we
now notify the client immediately of the failure.

Bug: 59585
Test: Added to bt-rfcomm-tests

Change-Id: I823dee7a3e739c5325aa716a9532ad3a32599b1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443861
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>