[oom][test] Disable flaky oom_test
The oom_test is flaking because a crash in driver_manager is causing the
device to reboot early. The test can sometimes trigger the OOM too soon,
catching the driver_manager in the middle of initialization, such that
allocations start failing with ZX_ERR_NO_MEMORY. This causes the
driver_manager to crash, and since it is a critical process, the device
reboots. This fails the test which specifically watches for an OOM
reboot.

Disable the test until it is rewritten to be more reliable.

Bug: 63505
Change-Id: Ia841438991c5c34b9773d1ec3e677bb7ab2972a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447296
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Testability-Review: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>