[display] Fake fuchsia.hardware.display.Provider service.
Refactors display driver testing code so that it can be repurposed
in an implementation of fuchsia.hardware.display.Provider which
vends a hermetic fake display controller, or rather: mostly hermetic.
This is because it uses the "global" sysmem.Allocator obtained from
the environment; this is so that E2E tests can be configured to
share the same Sysmem between the display controller, Vulkan,
and Scenic.

TEST=fx test display-core-unittests
TEST=fx test flatland_renderer_unittests
TEST=fx test root_presenter_unittests

Bug: 62388

Change-Id: I13fe7ff5cd4ad0e6b64c5286db260bbd97e0640a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427501
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Testability-Review: Randall Bosetti <rlb@google.com>