[elflib] Add explicit LLVM dependency
elflib is only using the ELF data structures, which are defined in LLVM
headers. It's important to provide a header-only LLVM target, as elflib
is built on both target and host, while LLVM is only available on host.

An alternative way is to avoid LLVM and use elf.h, which is not
available on macOS and requires certain config for compatibility.

Bug: 47403
Change-Id: I42807cb2ef43ea02d391c59534cbba360ebeaaf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446282
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Testability-Review: Casey Dahlin <sadmac@google.com>