[media][tools] Fix warnings in signal_generator
This CL fixes "implicit conversion leads to precision loss" warnings
in the signal_generator CLI utility, as mentioned in fxbug.dev/58162.
To the extent that this limits the size of a payload buffer (4GB) or
the theoretical maximum duration of a signal (292 yrs), this is now
code-commented, user-error-checked, and stated in the '--?' output.

Test: No behavior change; diagnostic CLI utility
BUG: 58162

Change-Id: If6bf667a90481fb00f467aeecc559c01d754926b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446597
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Payton Bradsky <pbradsky@google.com>
Testability-Review: Payton Bradsky <pbradsky@google.com>