[bt][avrcp] Ignore unknown feature bits
When a peer is discovered, the AVRCP SDP supported feature
bits have a number of unspecified bits, which are reserved
for additions.  These bits should be ignored, and the connection
should continue only using the features it is aware of.

Fixed: 45303, 46699, 46700
Test: IOPT/CL/AVRCP-CT/SFC/BV-03-I, IOPT/CL/AVRCP-TG/SFC/BV-04-I
      added service_from_attributes_ignores_unknown_features

Change-Id: I945db19eb40720336388edb55350d95685e54d0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447108
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Testability-Review: Ani Ramakrishnan <aniramakri@google.com>