[tests] Use console.shell output for shell start
The QEMU tests used to rely on a virtcon message to know when
the serial console was up and running. Virtcon isn't responsible
for the serial console, the line that it should be looking for
is:
   console.shell: enabled

This CL was reverted earlier to see if it helped fix a flake.
The flaking test has since been disabled and is being
rewritten.

Bug: 61528
TEST: Run the QEMU tests
Change-Id: I4feb09ae5a428b5369a738440867ba23594571d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447456
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>