[kernel][vm] Annotate locked state
This annotates the VmMapping state to indicate the locks needed to
manipulate it. Due to limitations in the clang capability analysis it
cannot determine that if two locks are needed to modify a member, then
holding either lock makes it safe to read. Therefore multiple accessor
functions are added for the two lock options.

Change-Id: Ib058001b065cbeadf99b9962d83de6d4fc218cb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447058
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>