[kernel][vm] Transform into hidden parent
This changes the VmCowPages::CreateClone behavior to instead of
creating a new hidden VmCowPages and moving state into it, convert
the existing VmCowPages to be hidden and create new children that
reference it.

The benefit to transforming the existing node is that moving the page
list between VmCowPages is expensive, as every page has to be walked to
update any backlinks. The tradeoff here is that all the children have
to be updated to have a new parent, but children are expected to be
either 0 or 2, which is much less than the number of pages that are
expected.

Change-Id: Id0904848c78d24377777a1d8f23a1c02d908f449
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437350
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>