[rust][fidl] Fix lifetime issue on array of tables
This CL changes a test in src/sys/pkg/bin/omaha-client/src/fidl.rs to
explicitly bind an array of FIDL tables to a variable before calling
iter() on it. Although the code compiles as is, without this change it
fails to compile when the `{}` are replaced by `::empty()`. This will be
done soon as part of enforcing source-compatible tables in Rust.

Bug: 7983
Test: CQ
Change-Id: If780e293f3177b20cf16484abc5389573840ead9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447577
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>