[pkg-cache] Maximize index iterator vector size
FIDL messages must fit within 64KiB, but it was possible for 32 index
entries to overflow a message (31 could not). This change increases the
maximum vector size for the package index iterator and modifies
pkg-cache to do the work to measure the size of elements in the
iterator, allowing it to fit as many items in a single FIDL message as
possible without overflowing it.

This change also fixes up the integration test for this iterator to not
need to construct a package for each index entry.

Change-Id: Iaff45a9591f528e7e734e6d7b232096c3a65f1a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447194
Commit-Queue: Kevin Wells <kevinwells@google.com>
API-Review: Dan Johnson <computerdruid@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Testability-Review: Ben Keller <galbanum@google.com>