[bt][rfcomm] Clean up credit flow mechanics
Updates the policy for including credits in frames. Based on behavior of
devices in the field, it's common to omit the credit octet when
the credits-to-be-issued is 0.

The Modem Status command is exchanged before user data is sent over the
DLC. GSM 7.10 Section 5.4.6.3.7 doesn't provide default values.
This CL updates the default value for the Data Valid (DV) bit to be set
based off of the behavior of devices seen in the field.

Moves channel.rs into the sessions mod as the SessionChannel is managed
by the RFCOMM Session.

Bug: 63415
Test: Updated bt-rfcomm-tests
Change-Id: If1bebfe9677bfa704497ab2e150281860775aec2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446277
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>