[rfc] Edge triggered async_wait
Documents a proposed change to zx_object_wait_async to
add a flag ZX_WAIT_ASYNC_EDGE which skips a check for
currently active signals and only queues a packet on the
port if a signal subsequently transitions from inactive
to active.

If any of the signals are active at the time of the
call of zx_object_wait_async, they must first transition
to the inactive state and then be asserted to generate
a packet on the port.

TEST: documentation only change.
Change-Id: Iffeca5718e904a5a5b8d59539a6e3005b5d4c8cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442834
Fuchsia-Auto-Submit: Steve Austin <steveaustin@google.com>
Commit-Queue: Steve Austin <steveaustin@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>