[test_spec] default to environments = basic_envs
By default (no gn args set) this is the same as the current default
of `[ emu_env ]`. The behavior will change for the case when the
gn arg `basic_names` is set.

We need this so that we can start running bringup tests based on the
contents of tests.json, rather than on specifying device_type in the
infrastructure. With this change, setting `basic_env_names` on a builder
will have a similar effect to the old setting of device_type.

Bug: 59963
Change-Id: I585cb187329f4f967dc08ca3c0b9a76c6b872cd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447458
Fuchsia-Auto-Submit: Gary Miguel <garymm@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>