[netemul] Add default route with specified NIC
... when using netstack_cfg.

Creating a default route without a specified NIC delegates the task of
finding the outgoing interface to the netstack. If a stack had 2
interfaces connected to the same network, the stack may only use
one of the interfaces as the outgoing interface. This will cause issues
when a stack may want to send packets through either of the interfaces
but only 1 will have the default route through it.

This change always specifies the NIC so that the default route goes
through the interface being configured.

Note that this is a limitation of the fuchsia.net.stack/Stack fidl which
does not allow clients to specify the outgoing interface when a route
depends on a next-hop node.

Change-Id: I8d3566b6432b03abe6ed299295e7edc53448f25d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447674
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>