[kernel][lib/arch] Introduce FakeCpuidIo
In this change, we update lib/arch to use BootCpuidIo (re-templating
accordingly) and introduce a in-test replacement for it: FakeCpuidIo.
FakeCpuidIo provides a simple interface for stubbing in dummy data and
is valid for use in the kernel as well as on host.

Bug: 61093
Change-Id: I426b4b4b99ac75c4db795d8f3da4a829c94e2433
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/446377
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>