[fidl][rust] Check handle type and rights on receive side
In a previous CL, FIDL Rust bindings were switched to use
zx_channel_read_etc which provides HandleInfo objects with rights and
type information. This CL checks that the rights and types are correct.

The main complication in implementing this was in making handle types
and rights from FIDL files available at decode time. Because there is no
way to embed this in the handle type itself, types that include handles
(e.g. structs, tables, etc) must include the rights information.

There is further complication with handles in method responses because
responses are represented by tuples (and possibly tuple of a Result of a
tuple), which doesn't have a way to include handle rights information in
the types. To support these cases, this CL provides a handle wrapper
type that is added to the type before decode and removed from the type
before returning a response.

Change-Id: I94be260b304500e514aa9b94bc76afc432c803b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444320
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>