[size_checker] minor refactor of map code
Uses the result of map lookup as the node to put information in
when computing file sizes.

This is posible because the maps in question store node pointers,
so we can use the pointer alias to modify the map entry.  It is not
always possible, but in this case it is, and it reduces the chattiness
of the code.

Bug: 63289
Change-Id: I992691432dcfc279aa37ad53ec9d4ee277733c7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447918
Testability-Review: Filip Filmar <fmil@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>