[wlan][sme] A BSS is only viable if get_protection() succeeds
When the client configuration and/or the provided credentials are
cannot be used to connect with a BSS, the BSS will be considered
incompatible or not viable. If a BSS is determined to be either, SME
will continue to try finding a BSS that is both. Practically, this
means even in the face of errors associated with setting up a
Protection for a BSS, SME will continue to make a best effort and try
to find a BSS for which it can create a Protection.

This change also modifies the SelecetNetworkFailure enum values.  The
NoCompatibleNetwork and CredentialError values are replaced by
IncompatibleConnectRequest and InternalProtectionError,
respectively. IncompatibleConnectRequest indicates that the connect
request received is malformed based on the scan results. In
particular, SME cannot connect to any BSSID in the scan results with
the connect request. InternalProtectionError indicates the connect
request is okay based on the scan results, but there is some other
reason why protection could not be derived for the network.

Test: Unit tests and manual testing
(1) fx test wlan-sme-tests wlan-common-tests wlan-hasher-tests \
connect-with-wrong-credential-type wlanstack-tests
(2) Manual test of connection to duplicate SSID network
Fixed: 62895

Change-Id: Id43018367b29ba3d73058ccf6ec341ace1b92766
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444975
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Reviewed-by: Sean Olson <seanolson@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Testability-Review: Dylan Swiggett <swiggett@google.com>