[settings] Serve values from light hardware configuration in light controller
The light API supports an optional hardware configuration that can be
specified through a json file. If the configuration is not present, the
light controller reads the underlying hardware API and reuses the names
the hardware returns, with each individual light getting it's own
LightGroup.

When a hardware configuration is specified, the product can define a
different mapping from physical lights to light groups, allowing logical
groupings of lights and specifying different options for them, such as
persistence.

While light controller already read a configuration previously, it did
not yet use it and still relied on the hardware lights. This CL will use
the light structure and names specified in the light hardware
configuration if it exists, falling back to the hardware otherwise.
Since the configuration is only used for tests at the moment, this CL
will not affect the behavior of the light API yet.

Fixed: 63554
Test: fx test setui_service_tests, locally specified an configuration
and verified it changed the output of the light API
MULTIPLY: setui_service_tests

Change-Id: I6618fe9f50cc4fbb2c51603f217c87aa99f28923
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447300
Commit-Queue: William Xiao <wxyz@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>