[storage] Remove barrier support from block driver.
This also fixes a shutdown race: once TxnEnd has been called and
pending_count_ is decremented, the instance of Server can be destroyed
so it wasn't safe to access data members after that.  Removing support
for barriers means that no members are now touched in TxnEnd after that
point.

This CL removes the use of atomics which were unlikely to help with
performance, and replaces them with a conventional mutex and condition.

There was no need for the additional message queue so that has been
removed; messages can be immediately queued with the lower layers.

Fixed: 41161
Change-Id: I207372b7ffa07d98252a37cf97822fa40b0c1016
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448080
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Chris Suter <csuter@google.com>