[image_pipe] Bind and use InterfacePtr on thread that uses it
fidl::InterfacePtr is a thread-hostile type so it must be bound and
destroyed on the same thread that it is used on. This updates the
ImagePipeSurfaceAsync class which uses a fidl::InterfacePtr on a
background thread to follow this guidance.

Bug: 63593
Change-Id: I4ed8488b75bc1a5437f714a4efcaf67196d0c7b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447916
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Testability-Review: John Bauman <jbauman@google.com>