[dart][build] Support package_root & non_dart_deps
This adds support for the package_root and non_dart_deps parameters to
the flutter_test_component and dart_test_component templates. These are
declared on the dart_library template and are needed to port some of the
FIDL tests over to these templates.

Change-Id: I8f0585e681879c22804928d10a4f269d53d5f463
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/447676
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>