[build] Fix scudo BUILD.gn file.
The binary equivalence between the libc.so binaries generated by the
Zircon and Fuchsia builds was broken (i.e. the files were no longer
byte identical). Investigation shows that this was due to a previous
CL [1], that changed the way the Scudo library is built in the
Fuchsia build, but not the Zircon build. Moreover, the CL was buggy
because it no longer compiled the src/flags.cpp source file with a
required compiler define and include path.

This CL fixes the issue, and ensures that BUILD.zircon.gn is now a
symlink into BUILD.gn to avoid this kind of mistake in the future.

For testing, I have manually verified that these two files are now
identical after the CL is applied:

  out/default/user.libc_x64/libc.so
  out/default.zircon/user-x64-clang.shlib/obj/system/ulib/c/libc.so

Bug: 60613
Testing: See above.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/446536

Change-Id: If014a00609dcc7dce3b503461bfbc724c9886fdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/448334
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Shai Barack <shayba@google.com>