[kernel] Timer: Convert active_cpu_ / cancel_ to ktl::atomic
Timer structs use active_cpu_ and cancel_ fields to track where a timer
callback is running and whether or not it was canceled. A timer can be
canceled from a different CPU than one that it's running / scheduled to
run on; it can be canceled while it is executing / as it is about to
start. Therefore the active_cpu_ / cancel_ fields need to be safe for
concurrent access.

The fields were marked volatile and arch::DeviceMemoryBarrier was used
after stores. Convert the fields to ktl::atomic, but keep the
DeviceMemoryBarriers in place at this point.

Most accesses can be converted to memory_order_relaxed; the cancel_
field in particular is set to indicate a timer is canceled. No dependent
accesses happen based on the cancel_ flag, so we do not need stronger
orders.

Tested:
1. k ut -r 100000 timer on qemu-kvm / x86
2. k ut -r 10000 timer on qemu-kvm w/ kCSAN

Bug: 47117 Use only ktl::atomic in the kernel
Bug: 61451 Kernelmode Concurrency / Thread Sanitizer for Zircon

Change-Id: I4ad0c1c17278c7bb848f9a85d082e7379fe3a027
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444274
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>